SDC-12512 Bug Fix
Pipelines (which were running before restart) are stuck in state `Starting`.
The root cause is that the snapshot name is always "default", so it will cause an Exception, and the state will not change.
I change the snapshot name, and add the uuid to the name. 
So it will work correctly. And the pipeline state will change to "Running".